refactor: image handling in components
