+++ apps/zeity/app/components/user/Menu.vue
@@ -29,7 +29,7 @@ const userMenu = computed(() => [
         ...organisations.value.map((organisation) => ({
             label: organisation.name,
             avatar: {
-                src: organisation.image || undefined,
+                src: getOrganisationImagePath(organisation),
                 alt: organisation.name,
             },
             slot: currentOrganisationId.value === organisation.id ? 'selectedOrg' as const : undefined,
@@ -48,10 +48,10 @@ const userMenu = computed(() => [
 
 <template>
     <UDropdownMenu :items="userMenu" size="xl">
-        <UAvatar :src="data?.user.image || undefined" :alt="data?.user.name" as="button"
+        <UAvatar :src="getUserImagePath(data?.user)" :alt="data?.user.name" as="button"
             class="ring ring-inset ring-primary/50 hover:bg-primary/10 focus:outline-none focus-visible:ring-2 focus-visible:ring-primary focus-visible:bg-primary/10" />
         <template #profile>
-            <UAvatar :src="data?.user.image || undefined" :alt="data?.user.name" size="xl" />
+            <UAvatar :src="getUserImagePath(data?.user)" :alt="data?.user.name" size="xl" />
             <div class="flex flex-col">
                 <h3 class="text-base font-semibold">
                     {{ data?.user.name }}
