+++ apps/zeity/app/utils/host.client.ts
@@ -1,0 +1,3 @@
+export function getHost() {
+  return document.location.origin;
+}
